// examples/convex/http/convex/_generated/api.ts
// Generated by @nimbus/codegen. Do not edit by hand.
import { makeActionReference, makeMutationReference, makeQueryReference } from "convex/browser";
import type { Doc, Id, JsonValue } from "./dataModel";

export const api = {
  messages: {
    list: makeQueryReference<{}, Doc<"messages">[]>("messages:list", "public"),
    byAuthor: makeQueryReference<{
  "author": string | null;
}, Doc<"messages">[]>("messages:byAuthor", "public"),
    maybeByAuthor: makeQueryReference<{
  "author": string | null;
}, ({
  "_id": Id<"messages">;
  "_creationTime": number;
  "_updateTime": number;
  "author": string;
  "body": string;
})[]>("messages:maybeByAuthor", "public"),
    byId: makeQueryReference<{
  "id": Id<"messages">;
}, Doc<"messages"> | null>("messages:byId", "public"),
    uniqueByAuthor: makeQueryReference<{
  "author": string;
}, Doc<"messages"> | null>("messages:uniqueByAuthor", "public"),
    exactByAuthorAndBody: makeQueryReference<{
  "author": string;
  "body": string;
}, Doc<"messages"> | null>("messages:exactByAuthorAndBody", "public"),
    send: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:send", "public"),
    sendViaAction: makeActionReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendViaAction", "public"),
    scheduleSend: makeMutationReference<{
  "author": string;
  "body": string;
  "delayMs": number;
}, string>("messages:scheduleSend", "public"),
    sendAndSchedule: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendAndSchedule", "public")
  }
};

export const internal = {
  messages: {
    sendInternal: makeMutationReference<{
  "author": string;
  "body": string;
}, Id<"messages">>("messages:sendInternal", "internal")
  }
};